Refactored text types & added text code documentation
