+++ src/components/ui/text/text.component.tsx
@@ -5,8 +5,15 @@ import {
 } from 'react-native';
 import React, { FC } from 'react';
 
-import styles from './text.component.styles';
 import { useTheme } from '@/hooks';
+import styles from './text.component.styles';
+
+import {
+  TextAppearance,
+  TextStatus,
+  TextVariants,
+  TextType,
+} from './text.types';
 
 export interface TextProps extends RNTextProps {
   variant?: TextVariants;
@@ -16,6 +23,39 @@ export interface TextProps extends RNTextProps {
   type?: TextType;
 }
 
+/**
+ * Basic text writing.
+ *
+ * @property {ReactText | ReactElement<TextProps>} children - String or number to be rendered as text.
+ * Also can be ReactElement<TextProps> - nested Text component.
+ *
+ * @property {string} variant - Can be `small`, `regular`, `large`, `title-small`, `title-regular` or `title-large`.
+ * Defaults to *regular*.
+ *
+ * @property {string} appearance - Can be `default`, `alternative` or `hint`.
+ * Use `alternative` for displaying light text on a dark content and vice versa.
+ * Use `hint` for giving user a hint on something.
+ *
+ * @property {string} status - Status of the component.
+ * Can be `basic`, `primary`, `success`, `info`, `warning`, `error` or `disabled`.
+ * Defaults to *basic*.
+ *
+ * @property {string} type - Type of the component.
+ * Can be `bold`, `light`, `italic`, `semi-bold` or `regular`.
+ * Defaults to *regular*.
+ *
+ * @property {TextProps} ...TextProps - Any props applied to Text component.
+ *
+ * @overview-example TextStyling
+ * Text can be styled with `style` property.
+ * In most cases this is redundant, if [custom theme is configured](guides/branding).
+ * ```
+ *
+ * import { Text } from '@/components/ui';
+ *
+ * <Text style={...}>Place your Text</Text>
+ * ```
+ */
 const Text: FC<TextProps> = ({ style, ...props }: TextProps) => {
   const theme = useTheme();
   const s = styles({ theme, props });
@@ -24,24 +64,3 @@ const Text: FC<TextProps> = ({ style, ...props }: TextProps) => {
 };
 
 export default Text;
-
-export type LiteralUnion<T extends U, U = string> = T | (U & {});
-
-export type TextType = LiteralUnion<
-  'bold' | 'light' | 'italic' | 'regular' | 'semi-bold'
->;
-
-export type TextStatus = LiteralUnion<
-  'basic' | 'primary' | 'success' | 'info' | 'warning' | 'error' | 'disabled'
->;
-
-export type TextAppearance = LiteralUnion<'default' | 'alternative' | 'hint'>;
-
-export type TextVariants = LiteralUnion<
-  | 'small'
-  | 'regular'
-  | 'large'
-  | 'title-small'
-  | 'title-regular'
-  | 'title-large'
->;
